requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.9.3
attrs==21.4.0
gurobipy==9.5.1
kiwisolver==1.4.2
numpy==1.22.3
ortools==9.3.10497
pddlpy==0.3.3
protobuf==3.19.4
python-dateutil==2.8.2
scipy==1.8.0
six==1.16.0
